requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
chromadb>=0.4.22
secure-smtplib>=0.1.1
